requirements.txt
crewai
openai
tavily
python-dotenv